flask_users_app/requirements.txt

flask==3.0.3
mysql-connector-python==9.0.0
